+++ src/linklayer/shaper/AS6802/CoRE4INET_TTShaper.h
@@ -565,8 +565,8 @@ bool TTShaper<TC>::isTransmissionAllowed(EtherFrame *message)
     SimTime sendTime = TC::outChannel->calculateDuration(message);
     //Don't know if that is right, but it works!
     sendTime += (INTERFRAME_GAP_BITS + ((PREAMBLE_BYTES + SFD_BYTES) * 8)) / TC::outChannel->getNominalDatarate();
-    unsigned long sendTicks = ceil((sendTime / getOscillator()->par("tick")).dbl());
-    unsigned long startTicks = ttBuffers.begin()->first;
+    uint64_t sendTicks = static_cast<uint64_t>(ceil((sendTime / getOscillator()->par("tick")).dbl()));
+    uint64_t startTicks = ttBuffers.begin()->first;
 
     if ((getTimer()->getTotalTicks() + sendTicks) >= startTicks)
     {
